Move test into storage
